requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
docx2pdf==0.1.8
dotenv==0.9.9
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
lxml==6.0.0
pdf2image==1.17.0
pdfkit==1.0.0
Pillow==9.5.0
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
pypdf==5.7.0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==310
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
wkhtmltopdf==0.2
